+++ test/fake_store_api_client_test.dart
@@ -77,5 +77,4 @@ void main() {
       expect(failure.message, customMessage);
     });
   });
-
 }